Refactor and add link to get ticket button
