+++ src/content/eventInfo.js
@@ -1,0 +1,5 @@
+const EventInfo = {
+  ticketUrl: "https://google.com",
+};
+
+export default EventInfo;
